requirements.txt
urllib3
winshell
imutils
numpy
opencv_python
Pillow
pypiwin32
ttkthemes
pyautogui
requests
beautifulsoup4
whatsmyip
pynput
pytesseract